mplayer svn: bump SRCREV, refresh patches

+++ recipes/mplayer/mplayer_svn.bb
@@ -10,22 +10,25 @@ LICENSE = "GPL"
 SRC_URI = "svn://svn.mplayerhq.hu/mplayer;module=trunk \
 	   file://makefile-nostrip-svn.patch \
 	   file://mplayer-arm-pld.patch \
-	   file://mplayer-lavc-arm.patch \
+	   file://mplayer-lavc-arm.patch;maxrev=30166 \
        file://fix-exp.diff;maxrev=30291 \
 	   file://fix-addrinfo.patch;maxrev=30302 \
        file://fix-avconfig.diff;maxrev=30376 \
 	   file://fix-emu_qtx_api.diff;maxrev=30165 \
-       file://codecs_conf-VP8.diff;striplevel=0 \
-       file://demux_mkv-V_VP8__webm_doctype.diff;striplevel=0 \
-       file://configure-libvpx_test.diff \
+       file://codecs_conf-VP8.diff;striplevel=0;maxrev=30166 \
+       file://demux_mkv-V_VP8__webm_doctype.diff;striplevel=0;maxrev=30166 \
+       file://configure-libvpx_test.diff;maxrev=30166 \
        file://vofw-swscale.diff \
+       file://offset.patch;minrev=32735 \
+       file://0001-MPlayer-FFmpeg-VP8-encode-decode-patches-using-libvp.patch;minrev=32735 \
 "
 
-SRCREV = "30165"
+SRCREV = "32735"
 SRC_URI_append_armv7a = " \
-		file://omapfb.patch \
-	   file://vo_omapfb.c \
-	   file://yuv.S \
+       file://omapfb.patch;maxrev=30166 \
+       file://vo_omapfb.c \
+       file://yuv.S \
+       file://0001-omapfb.patch;minrev=30166 \
 	  "
 
 # This is required for the collie machine only as all stacks in that
@@ -71,7 +74,6 @@ EXTRA_OECONF = " \
 	--enable-mencoder \
 	--disable-gui \
 	--enable-largefiles \
-	--disable-linux-devfs \
 	--disable-lirc \
 	--disable-lircc \
 	--disable-joystick \
@@ -81,8 +83,8 @@ EXTRA_OECONF = " \
 	--enable-tv-v4l1 \	 
 	--enable-tv-v4l2 \
 	--disable-tv-bsdbt848 \
+	--enable-networking \
 	--enable-rtc \
-	--enable-network \
 	--disable-smb \
 	--enable-live \
 	--disable-dvdnav \
@@ -109,10 +111,7 @@ EXTRA_OECONF = " \
 	--disable-xvid \
 	--disable-x264 \
 	\
-	--disable-libavutil_so \
-	--disable-libavcodec_so \
-	--disable-libavformat_so \
-	--disable-libpostproc_so \
+	--disable-ffmpeg_so \
 	\
 	--enable-tremor-low \
 	\
@@ -140,8 +139,7 @@ EXTRA_OECONF = " \
 	--disable-directx \
 	--disable-dxr2 \
 	--disable-dxr3 \
-	--disable-dvb \
-	--enable-dvbhead \
+	--enable-dvb \
 	--disable-mga \
 	--disable-xmga \
 	--enable-xv \
@@ -177,9 +175,9 @@ EXTRA_OECONF = " \
 	--disable-sunaudio \
 	--disable-win32waveout \
 	--enable-select \
-	--enable-libvpx-vp8-lavc \
+	--enable-libvpx-lavc \
 	\
-	--extra-libs=' -lBasicUsageEnvironment -lUsageEnvironment -lgroupsock -lliveMedia -lstdc++' \
+	--extra-libs=' -lvpx -lBasicUsageEnvironment -lUsageEnvironment -lgroupsock -lliveMedia -lstdc++' \
     --enable-protocol='file_protocol pipe_protocol http_protocol rtmp_protocol tcp_protocol udp_protocol' \
 "
 
@@ -202,19 +200,26 @@ EXTRA_OECONF_append = " ${@base_contains('MACHINE_FEATURES', 'iwmmxt', ' --enabl
 EXTRA_OECONF_append = " ${@base_contains('MACHINE_FEATURES', 'x86', '--enable-runtime-cpudetection', '',d)} "
 
 FULL_OPTIMIZATION = "-fexpensive-optimizations -fomit-frame-pointer -frename-registers -O4 -ffast-math"
-FULL_OPTIMIZATION_armv7a = "-fexpensive-optimizations  -ftree-vectorize -fomit-frame-pointer -O4 -ffast-math"
+FULL_OPTIMIZATION_armv7a = "-fno-tree-vectorize -fomit-frame-pointer -O4 -frename-registers -ffast-math"
 BUILD_OPTIMIZATION = "${FULL_OPTIMIZATION}"
 # FIXME: Temporarily disable debugging to work-around http://gcc.gnu.org/bugzilla/show_bug.cgi?id=37987
 DEBUG_OPTIMIZATION_spitz = "-O -fomit-frame-pointer -g"
 DEBUG_OPTIMIZATION_akita = "-O -fomit-frame-pointer -g"
 
 do_configure_prepend_armv7a() {
-	cp ${WORKDIR}/yuv.S ${S}/libvo
- 	cp ${WORKDIR}/vo_omapfb.c ${S}/libvo
+	if [ -e ${S}/libvo/yuv.S ] ; then
+		echo "files already present"
+	else
+		cp ${WORKDIR}/yuv.S ${S}/libvo
+	 	cp ${WORKDIR}/vo_omapfb.c ${S}/libvo
+	fi
 	cp ${STAGING_KERNEL_DIR}/arch/arm/plat-omap/include/mach/omapfb.h ${S}/libvo/omapfb.h || true
  	cp ${STAGING_KERNEL_DIR}/include/asm-arm/arch-omap/omapfb.h ${S}/libvo/omapfb.h || true
 	cp ${STAGING_KERNEL_DIR}/include/linux/omapfb.h ${S}/libvo/omapfb.h || true
  	sed -e 's/__user//g' -i ${S}/libvo/omapfb.h || true
+
+	# Don't use hardfp args when using softfp
+	sed -i -e 's:if HAVE_VFP_ARGS:ifdef __ARM_PCS_VFP:' ${S}/ffmpeg/libavcodec/arm/asm.S
 }
 
 CFLAGS_append = " -I${S}/libdvdread4 "
